Makeing project compilable.

+++ SubmarineSimulator/SubmarineSimulator/FoodSupply.cs
@@ -6,7 +6,7 @@ namespace SubmarineSimulator
 {
     public class FoodSupply : IStorable
     {
-        public int Amount { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
+        public double Amount { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
 
         public void HowMuchLeft()
         {
