+++ SubmarineSimulator/SubmarineSimulator/OxygenSupply.cs
@@ -19,7 +19,7 @@ namespace SubmarineSimulator
             }
         }
 
-        public int Amount { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
+        public double Amount { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
 
         public void HowMuchLeft()
         {
