+++ SubmarineSimulator/SubmarineSimulator/Waste.cs
@@ -13,7 +13,7 @@ namespace SubmarineSimulator
             throw new System.NotImplementedException();
         }
 
-        public int Amount { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
+        public double Amount { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
 
         public void HowMuchLeft()
         {
